Added RoomsControllerTests

+++ SensorService/Controllers/RoomsController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Common.Repository;
 using Domain.Entities;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Environment = Domain.Entities.Environment;
 
@@ -26,6 +27,11 @@ namespace SensorService.Controllers
         public async Task<IActionResult> Get(int id)
         {
             var room = await this.homeRepository.GetRoom(id);
+
+            if (room == null)
+            {
+                return NotFound();
+            }
             return Ok(room);
         }
 
@@ -33,7 +39,11 @@ namespace SensorService.Controllers
         [HttpPut]
         public async Task<IActionResult> Put([FromBody]Room room)
         {
-            await this.homeRepository.EditRoom(room);
+            var success = await this.homeRepository.EditRoom(room);
+            if (!success)
+            {
+                return BadRequest();
+            }
             return Ok(room);
         }
 
@@ -41,6 +51,10 @@ namespace SensorService.Controllers
         public async Task<IActionResult> GetBySensor([FromQuery(Name = "sensorId")]int sensorId)
         {
             var room = await this.homeRepository.GetRoomBySensorId(sensorId);
+            if (room == null)
+            {
+                return NotFound();
+            }
             return Ok(room);
         }
 
@@ -49,7 +63,15 @@ namespace SensorService.Controllers
         public async Task<IActionResult> Delete(int id)
         {
             var room = await this.homeRepository.GetRoom(id);
-            await this.homeRepository.DeleteRoom(id);
+            if (room == null)
+            {
+                return NotFound();
+            }
+            var success = await this.homeRepository.DeleteRoom(id);
+            if (!success)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
 
             return Ok(room);
         }
@@ -58,6 +80,10 @@ namespace SensorService.Controllers
         public async Task<IActionResult> GetSensors(int id)
         {
             var room = await this.homeRepository.GetRoom(id);
+            if (room == null)
+            {
+                return NotFound();
+            }
             var sensors = await this.homeRepository.GetSensors(room);
             return Ok(sensors);
         }
@@ -65,13 +91,35 @@ namespace SensorService.Controllers
         [HttpPost("{id}/sensor")]
         public async Task<IActionResult> PostSensor(int id, [FromBody]Sensor sensor)
         {
-            await this.homeRepository.AddSensor(id, sensor);
+            var room = await this.homeRepository.GetRoom(id);
+            if (room == null)
+            {
+                return NotFound();
+            }
+            var insertedSensor = await this.homeRepository.AddSensor(id, sensor);
+            if (insertedSensor == null)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
+
             return Ok(sensor);
         }
 
         [HttpGet("{id}/environment")]
         public async Task<IActionResult> GetEnvironment(int id, [FromQuery]DateTime startDate, [FromQuery]DateTime endDate)
         {
+            var invalidDate = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
+            if (!ModelState.IsValid || startDate.Equals(endDate) || startDate.Equals(invalidDate) || endDate.Equals(invalidDate))
+            {
+                return BadRequest();
+            }
+
+            var room = await this.homeRepository.GetRoom(id);
+            if (room == null)
+            {
+                return NotFound();
+            }
+
             var envs = await this.homeRepository.GetEnvironmentReadings(id, startDate, endDate);
             return Ok(envs);
         }
@@ -79,7 +127,23 @@ namespace SensorService.Controllers
         [HttpPost("{id}/sensor/{sensorId}/environment")]
         public async Task<IActionResult> PostEnvironment(int id, int sensorId, [FromBody]Environment environment)
         {
+            var room = await this.homeRepository.GetRoom(id);
+            if (room == null)
+            {
+                return NotFound();
+            }
+
+            var sensor = await this.homeRepository.GetSensor(sensorId);
+            if (sensor == null)
+            {
+                return NotFound();
+            }
+
             var insertedEnvironment = await this.homeRepository.AddEnvironmentReading(id, sensorId, environment);
+            if (insertedEnvironment == null)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
             return Ok(insertedEnvironment);
         }
     }
